refactor: 정수 범위 검증 공통 메소드 추출 후 이전
